clean up unit tests
